Make viewController log Rect using Factory

+++ DrawingApp/DrawingApp/Model/Model/Factory.swift
@@ -10,8 +10,8 @@ import Foundation
 class Factory {
     func createRandomSquare() -> Square? {
         let (width, height) = (Double(150), Double(120))
-        let X = Double.random(in: 10...500)
-        let Y = Double.random(in: 10...500)
+        let X = Double(Int.random(in: 10...500))
+        let Y = Double(Int.random(in: 10...500))
         guard let color = Color(r: Int.random(in: 0...255),
                                 g: Int.random(in: 0...255),
                                 b: Int.random(in: 0...255),
